+++ src/constants/index.ts
@@ -1,0 +1,37 @@
+// Application limits
+export const MAX_IMAGES = 50;
+
+// Unit conversions
+export const CM_PER_INCH = 2.54;
+export const INCHES_PER_FOOT = 12;
+
+// Color palette options
+export const COLOR_OPTIONS = [
+  '#F9A826', // Orange
+  '#F26D6D', // Pink/Red
+  '#42CAFD', // Light Blue
+  '#7747FF', // Purple
+  '#FD5EB3', // Hot Pink
+  '#743886', // Dark Purple
+];
+
+// Silhouette paths
+export const MALE_SILHOUETTE_SVG = '/images/male.svg';
+export const FEMALE_SILHOUETTE_SVG = '/images/female.svg';
+
+// Aspect ratios that match the silhouettes
+export const MALE_ASPECT_RATIO = 0.4;
+export const FEMALE_ASPECT_RATIO = 0.4;
+
+// ImageComparer constants
+export const SCALE_TOP_FACTOR = 1.1;
+export const MAJOR_INTERVALS = 10;
+export const FIGURE_LABEL_OFFSET_Y = -1;
+export const IMAGE_GAP_PX = 16;
+
+// Drag functionality constants
+export const DRAG_THRESHOLD_PX = 20;
+export const LONG_PRESS_DURATION = 100;
+
+// Calculation epsilon
+export const EPSILON = 1e-6;
